requirements.txt
# Core dependencies
openvino-genai>=2025.1.0
openvino>=2025.1.0
openvino-tokenizers>=2025.1.0

# HuggingFace for model download
huggingface-hub>=0.20.0

# Environment configuration
python-dotenv>=1.0.0
